edit detailItem page
remove delete function from homePage and add it to deatilItemPage
